+++ modules/network/subnets.tf
@@ -105,3 +105,27 @@ resource "aws_subnet" "elasticsearch_2" {
     PlatformInstanceID = "${var.platform_instance_id}"
   }
 }
+
+resource "aws_subnet" "faces_1" {
+  vpc_id            = "${aws_vpc.main.id}"
+  cidr_block        = "${var.cidr_subnet_faces_1}"
+  availability_zone = "${var.az1}"
+
+  tags {
+    Name               = "GrayMetaPlatform-${var.platform_instance_id}-Faces1"
+    Application        = "GrayMetaPlatform"
+    PlatformInstanceID = "${var.platform_instance_id}"
+  }
+}
+
+resource "aws_subnet" "faces_2" {
+  vpc_id            = "${aws_vpc.main.id}"
+  cidr_block        = "${var.cidr_subnet_faces_2}"
+  availability_zone = "${var.az2}"
+
+  tags {
+    Name               = "GrayMetaPlatform-${var.platform_instance_id}-Faces2"
+    Application        = "GrayMetaPlatform"
+    PlatformInstanceID = "${var.platform_instance_id}"
+  }
+}