+++ modules/services/security_groups.tf
@@ -63,7 +63,7 @@ resource "aws_security_group" "services" {
   }
 }
 
-resource "aws_security_group" "services_alb_security_group" {
+resource "aws_security_group" "services_alb" {
   description = "Services ALB Security Group"
   vpc_id      = "${data.aws_subnet.subnet_1.vpc_id}"
 
@@ -85,21 +85,21 @@ resource "aws_security_group" "services_alb_security_group" {
     from_port   = 443
     to_port     = 443
     protocol    = "tcp"
-    cidr_blocks = ["${var.ecs_nat_ip}"]
+    cidr_blocks = ["${var.az1_nat_ip}"]
   }
 
   ingress {
     from_port   = 443
     to_port     = 443
     protocol    = "tcp"
-    cidr_blocks = ["${var.services_nat_ip}"]
+    cidr_blocks = ["${var.az2_nat_ip}"]
   }
 
   ingress {
     from_port   = 8445
     to_port     = 8445
     protocol    = "tcp"
-    cidr_blocks = ["${var.ecs_nat_ip}"]
+    cidr_blocks = ["${var.az1_nat_ip}"]
   }
 
   egress {